wpilib.testing: Extract shared pytest plugin test helpers
Move robot module generation and plugin configuration into a reusable
helper module for the existing suite and upcoming ordering tests.
